+++ nagra/update.py
@@ -8,7 +8,7 @@ from nagra.exceptions import ValidationError
 from nagra.transaction import Transaction
 from nagra.upsert import Upsert
 from nagra.writer import WriterMixin
-from nagra.utils import snake_to_pascal, get_table_from_dataclass, iter_dataclass_cols
+from nagra.utils import get_table_from_dataclass, iter_dataclass_cols
 
 
 if TYPE_CHECKING: